Refactor system hostname
